add RelayStoreUtils
